Added ModificationTrackerBaseEntity
